metric card ajax with update time

+++ resources/js/backend/app.js
@@ -16,16 +16,18 @@ const vue = new Vue({
         alerts: []
     },
     mounted() {
-        this.$http.post(`${laravel.root_url}/admin/vstack/notifications/${laravel.user.code}`, {}).then(res => {
-            res = res.data
-            if (res.notifications) this.alerts = res.notifications
-        })
-        if (laravel.user && laravel.pusher_key) {
-            let route = `App.User.${laravel.user.id}`
-            Echo.private(route).notification(n => {
-                this.$message({ showClose: true, message: n.message, type: n._type })
-                this.$http.delete(`${laravel.root_url}/admin/vstack/notifications/${laravel.user.code}/${n.id}/destroy`, {})
+        if (laravel.user) {
+            this.$http.post(`${laravel.root_url}/admin/vstack/notifications/${laravel.user.code}`, {}).then(res => {
+                res = res.data
+                if (res.notifications) this.alerts = res.notifications
             })
+            if (laravel.user && laravel.pusher_key) {
+                let route = `App.User.${laravel.user.id}`
+                Echo.private(route).notification(n => {
+                    this.$message({ showClose: true, message: n.message, type: n._type })
+                    this.$http.delete(`${laravel.root_url}/admin/vstack/notifications/${laravel.user.code}/${n.id}/destroy`, {})
+                })
+            }
         }
     }
 })
